Fix tons of stylecop & documentation defects.
